feat(countries): se genero un archivo con paises, que ya se pueden consumir en el front provisoriamente
